+++ java/src/main/java/com/techelevator/dao/RecipeDao.java
@@ -5,15 +5,15 @@ import com.techelevator.model.Recipe;
 import java.util.List;
 
 public interface RecipeDao {
-    Recipe getRecipe(int id);
+    Recipe getRecipe(long id);
 
     List<Recipe> listRecipe();
 
-    List<Recipe> listRecipesByIngredient(int ingredientId);
+    List<Recipe> listRecipesByIngredient(long ingredientId);
 
     List<Recipe> listRecipesByType(String type);
 
     boolean addRecipe(Recipe recipe);
 
-    boolean deleteRecipe(int id);
+    boolean deleteRecipe(long id);
 }
